+++ src/index.ts
@@ -2,7 +2,7 @@ export * from './messages/formatters';
 
 export * as SlashCommandAssertions from './interactions/slashCommands/Assertions';
 export * from './interactions/slashCommands/SlashCommandBuilder';
-export * from './interactions/slashCommands/SlashCommandSubCommands';
+export * from './interactions/slashCommands/SlashCommandSubcommands';
 export * from './interactions/slashCommands/options/boolean';
 export * from './interactions/slashCommands/options/channel';
 export * from './interactions/slashCommands/options/integer';